Make ITheme/IPartial to be identical with Theme/PartialTheme typing.
