+++ packages/theme/etc/theme.api.md
@@ -58,13 +58,15 @@ export namespace CommunicationColors {
     tint40 = "#eff6fc";
 }
 
-// @public (undocumented)
+// @public
+export type ComponentsStyles = {
+    [componentName: string]: ComponentStyles;
+};
+
+// @public
 export interface ComponentStyles {
-    // (undocumented)
-    [componentName: string]: {
-        styles?: IStyleFunctionOrObject<any, any>;
-        variants?: Variants;
-    };
+    styles?: IStyleFunctionOrObject<any, any>;
+    variants?: Variants;
 }
 
 // @public (undocumented)
@@ -383,20 +385,7 @@ export interface IPalette {
 }
 
 // @public (undocumented)
-export type IPartialTheme = {
-    palette?: Partial<IPalette>;
-    fonts?: Partial<IFontStyles>;
-    defaultFontStyle?: IRawStyle;
-    semanticColors?: Partial<ISemanticColors>;
-    isInverted?: boolean;
-    disableGlobalClassNames?: boolean;
-    rtl?: boolean;
-    spacing?: Partial<ISpacing>;
-    effects?: Partial<IEffects>;
-    schemes?: {
-        [P in ISchemeNames]?: IScheme;
-    };
-};
+export type IPartialTheme = PartialTheme;
 
 // @public (undocumented)
 export interface IScheme {
@@ -556,12 +545,7 @@ export interface ISpacing {
 }
 
 // @public (undocumented)
-export interface ITheme extends IScheme {
-    // @internal
-    schemes?: {
-        [P in ISchemeNames]?: IScheme;
-    };
-}
+export type ITheme = Theme;
 
 // @public (undocumented)
 export namespace LocalizedFontFamilies {
@@ -733,12 +717,35 @@ export namespace NeutralColors {
 }
 
 // @public
-export interface PartialTheme extends IPartialTheme {
+export interface PartialTheme {
     // (undocumented)
-    components?: ComponentStyles;
+    components?: ComponentsStyles;
+    defaultFontStyle?: IRawStyle;
+    // (undocumented)
+    disableGlobalClassNames?: boolean;
+    // (undocumented)
+    effects?: Partial<IEffects>;
+    // (undocumented)
+    fonts?: Partial<IFontStyles>;
+    // (undocumented)
+    isInverted?: boolean;
+    // (undocumented)
+    palette?: Partial<IPalette>;
+    // (undocumented)
+    rtl?: boolean;
+    // @internal
+    schemes?: {
+        [P in ISchemeNames]?: IScheme;
+    };
+    // (undocumented)
+    semanticColors?: Partial<ISemanticColors>;
+    // Warning: (ae-incompatible-release-tags) The symbol "spacing" is marked as @public, but its signature references "ISpacing" which is marked as @internal
+    //
+    // (undocumented)
+    spacing?: Partial<ISpacing>;
     // (undocumented)
     stylesheets?: string[];
-    // (undocumented)
+    // @internal
     tokens?: RecursivePartial<Tokens>;
 }
 
@@ -826,12 +833,14 @@ export namespace SharedColors {
 export type SizeValue = 'smallest' | 'smaller' | 'small' | 'medium' | 'large' | 'larger' | 'largest';
 
 // @public
-export interface Theme extends ITheme {
-    // (undocumented)
-    components?: ComponentStyles;
-    // (undocumented)
+export interface Theme extends IScheme {
+    components?: ComponentsStyles;
+    // @internal
     id?: string;
-    // (undocumented)
+    // @internal
+    schemes?: {
+        [P in ISchemeNames]?: IScheme;
+    };
     stylesheets?: string[];
     // @internal
     tokens?: RecursivePartial<Tokens>;
@@ -854,14 +863,9 @@ export type TokenSetType = {
     [key: string]: TokenSetType | string | number | undefined;
 };
 
-// @public (undocumented)
+// @public
 export type Variants = Record<string, any>;
 
-
-// Warnings were encountered during analysis:
-//
-// lib/types/ITheme.d.ts:70:5 - (ae-incompatible-release-tags) The symbol "spacing" is marked as @public, but its signature references "ISpacing" which is marked as @internal
-// lib/types/ITheme.d.ts:72:5 - (ae-incompatible-release-tags) The symbol "schemes" is marked as @public, but its signature references "ISchemeNames" which is marked as @internal
 
 // (No @packageDocumentation comment for this package)
